refactor(KERN): remove legacy design Tiles
