--- app/components/formElements/tile/TileRadio.tsx
@@ -1,85 +1,0 @@
-import { useField } from "@rvf/react-router";
-import TileTag, { type TileDescriptionProps } from "./TileTag";
-import Image, { type ImageProps } from "../../common/Image";
-
-const IMAGE_HEIGHT = 32;
-const IMAGE_WIDTH = 32;
-
-export type TileOptions = Readonly<
-  {
-    value: string;
-    description?: string;
-    title?: string;
-    image?: ImageProps;
-  } & TileDescriptionProps
->;
-
-type TileProps = TileOptions &
-  Readonly<{
-    name: string;
-    onClick: () => void;
-    ref: React.Ref<HTMLInputElement>;
-  }>;
-
-function TileRadio({
-  name,
-  description,
-  value,
-  title,
-  image,
-  tagDescription,
-  onClick,
-  ref,
-}: TileProps) {
-  const field = useField(name);
-  const id = `${name}-${value}`;
-  const errorId = `${name}-error`;
-  const descriptionId = description ? `${value}-description` : undefined;
-  const ariaDescribedBy = field.error() ? errorId : descriptionId;
-
-  return (
-    <div className="ds-tile-radio-group rounded-lg border-2 border-[#B3C9D6] bg-white">
-      <label
-        className="flex flex-row items-center cursor-pointer touch-manipulation p-24 h-full"
-        htmlFor={id}
-        aria-label={title}
-      >
-        <input
-          {...field.getInputProps({ type: "radio", id, value })}
-          className="ds-radio forced-colors:outline-solid forced-colors:border-[ButtonText]"
-          name={name}
-          type="radio"
-          aria-describedby={ariaDescribedBy}
-          onClick={onClick}
-          ref={ref}
-        />
-        <div className="h-full pl-24 flex flex-col gap-y-8 max-w-[calc(100%-1.5rem)]">
-          <div className="flex flex-row justify-between">
-            {image && (
-              <Image
-                {...image}
-                width={IMAGE_WIDTH}
-                height={IMAGE_HEIGHT}
-                ariaHidden={true}
-              />
-            )}
-            <TileTag tagDescription={tagDescription} />
-          </div>
-          <div>
-            <p className="ds-label-01-bold">{title}</p>
-            {description && (
-              <p
-                id={descriptionId}
-                className="ds-body-01-reg whitespace-pre-line break-words"
-              >
-                {description}
-              </p>
-            )}
-          </div>
-        </div>
-      </label>
-    </div>
-  );
-}
-
-export default TileRadio;
