+++ stories/form/Tile.stories.tsx
@@ -1,11 +1,11 @@
 import { reactRouterFormContext } from "~/../.storybook/reactRouterFormContext";
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import KernTile from "~/components/kern/formElements/tile/KernTile";
 import { Grid } from "~/components/layout/grid/Grid";
 import { GridItem } from "~/components/layout/grid/GridItem";
 import { GridSection } from "~/components/layout/grid/GridSection";
 import img1Url from "../assets/nichtbefoerderung.svg?no-inline";
 import img2Url from "../assets/verspaetung.svg?no-inline";
+import KernTile from "~/components/formElements/tile/KernTile";
 
 const meta = {
   title: "form/Tile",
